Extracting SoftKeyboardUtil from Util
